LayoutTests/http/tests/security/contentSecurityPolicy/worker-set-timeout-blocked-expected.txt

CONSOLE MESSAGE: Refused to execute a script because 'unsafe-eval' does not appear in the script-src directive of the Content Security Policy.
ALERT: setTimeout blocked
